feat: Add a home page for authenticated user Change the current home page to a landing page Login function and GuestRoute navigate user to /home
