+++ learnhub-ui/src/App.tsx
@@ -3,28 +3,26 @@ import { useEffect, useState } from "react";
 import { Routes, Route } from "react-router-dom";
 import {
     NotFound,
-    Home,
+    Landing,
     Login,
+    ManagerLogin,
     Register,
     Unauthorized,
     ForgotPassword,
     ResetPassword,
-    StudentDashboard,
-    TeacherDashboard,
     CourseList,
-    TeacherDetail
+    TeacherDetail,
+    ContactUs,
+    Home,
+    Mailbox
 } from "./pages";
-import { ManagerDashboard, ManagerLogin } from "./pages/manager";
 import GuestRoute from "./routers/GuestRoute";
 import ProtectedRoute from "./routers/ProtectedRoute";
 import Dummy from "./pages/Dummy";
-import AuthProvider from "./hooks/useAuth";
 import { ToastContainer } from "react-toastify";
 import { UserRole } from "./types/Account";
-import ContactUs from "./pages/ContactUs";
-import { DashboardLayout } from "./layouts";
-import Mailbox from "./pages/manager/teacher/Mailbox";
 import UserProfile from "./pages/UserProfile";
+import AuthProvider from "./hooks/useAuth";
 
 export default function App() {
     const [isLoading, setLoading] = useState(true);
@@ -48,22 +46,19 @@ export default function App() {
     return (
         <AuthProvider>
             <Routes>
-                <Route path="/" element={<Home />} />
-                <Route
-                    path="/test"
-                    element={
-                        <DashboardLayout>
-                            <></>
-                        </DashboardLayout>
-                    }
-                />
-
+                <Route element={<ProtectedRoute />}>
+                    <Route path="/dummy" element={<Dummy />} />
+                </Route>
+                {/* Public Routes */}
+                <Route path="/" element={<Landing />} />
+                <Route path="/contact" element={<ContactUs />} />
                 <Route path="/activate/:token" element={<Login />} />
+                <Route path="/courses" element={<CourseList />} />
+                <Route path="/teacher/:id" element={<TeacherDetail />}></Route>
+
+                {/* Routes for unathenticated users */}
                 <Route element={<GuestRoute />}>
                     <Route path="/login" element={<Login />} />
-                </Route>
-                <Route element={<ProtectedRoute />}>
-                    <Route path="/user-profile" element={<UserProfile />}></Route>
                 </Route>
                 <Route element={<GuestRoute />}>
                     <Route path="/manager/login" element={<ManagerLogin />} />
@@ -77,24 +72,19 @@ export default function App() {
                 <Route element={<GuestRoute />}>
                     <Route path="/reset-password/:token" element={<ResetPassword />} />
                 </Route>
-                <Route path="/teacher-profile/:id" element={<TeacherDetail />}></Route>
-                <Route path="/contact" element={<ContactUs />} />
+
+                {/* Routes for authenticated users */}
                 <Route element={<ProtectedRoute />}>
-                    <Route path="/dummy" element={<Dummy />} />
+                    <Route path="/home" element={<Home />} />
                 </Route>
-                <Route element={<ProtectedRoute roles={[UserRole.STUDENT]} />}>
-                    <Route path="/learning" element={<StudentDashboard />} />
+                <Route element={<ProtectedRoute />}>
+                    <Route path="/profile" element={<UserProfile />}></Route>
                 </Route>
-                <Route element={<ProtectedRoute roles={[UserRole.TEACHER]} />}>
-                    <Route path="/dashboard" element={<TeacherDashboard />} />
-                </Route>
-                <Route element={<ProtectedRoute roles={[UserRole.TEACHER_MANAGER, UserRole.COURSE_MANAGER]} />}>
-                    <Route path="/manager/dashboard" element={<ManagerDashboard />} />
-                </Route>
-                <Route path="/courses" element={<CourseList />} />
                 <Route element={<ProtectedRoute roles={[UserRole.TEACHER_MANAGER]} />}>
                     <Route path="/manager/mailbox" element={<Mailbox />} />
                 </Route>
+
+                {/* Error Boundary */}
                 <Route path="/unauthorized" element={<Unauthorized />} />
                 <Route path="/*" element={<NotFound />} />
             </Routes>
