+++ learnhub-ui/src/pages/StudentHome.tsx
@@ -1,0 +1,9 @@
+import { HomeLayout } from "../layouts";
+
+export default function StudentHome() {
+    return (
+        <HomeLayout>
+            <h1>This is Student Home</h1>
+        </HomeLayout>
+    );
+}
