+++ learnhub-ui/src/pages/TeacherHome.tsx
@@ -1,0 +1,9 @@
+import { HomeLayout } from "../layouts";
+
+export default function TeacherHome() {
+    return (
+        <HomeLayout>
+            <h1>This is Teacher Home</h1>
+        </HomeLayout>
+    );
+}
