+++ learnhub-ui/src/pages/TeacherManagerHome.tsx
@@ -1,0 +1,9 @@
+import { HomeLayout } from "../layouts";
+
+export default function TeacherManagerHome() {
+    return (
+        <HomeLayout>
+            <h1>This is Teacher Manager home</h1>
+        </HomeLayout>
+    );
+}
