+++ learnhub-ui/src/routers/GuestRoute.tsx
@@ -4,7 +4,7 @@ import { useAuth } from "../hooks/useAuth";
 export default function GuestRoute() {
     const { account } = useAuth();
     if (account) {
-        return <Navigate to="/" replace />;
+        return <Navigate to="/home" replace />;
     }
     return <Outlet />;
 }
